+++ CHANGELOG.md
@@ -7,10 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 ### Changed
+- Maximum datagram size in bytes for unix datagram generator is 8,192.
 - Increased maximum DogStatsD context limit from 100k to 1M
 ### Fixed
 - The capture manager will no longer panic if recording a capture and checking for a shutdown combined takes longer than one second.
 - A shutdown race was partially fixed in the capture manager which could result in truncated (invalid) json capture files.
+- Unix datagram generator will not longer 'sheer' blocks across datagrams.
 
 ## [0.20.10]
 ### Added