+++ CHANGELOG.md
@@ -9,10 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Maximum datagram size in bytes for unix datagram generator is 8,192.
 - Increased maximum DogStatsD context limit from 100k to 1M
+- Procfs generator now generates process names as long as 254 characters, up from 253.
 ### Fixed
 - The capture manager will no longer panic if recording a capture and checking for a shutdown combined takes longer than one second.
 - A shutdown race was partially fixed in the capture manager which could result in truncated (invalid) json capture files.
 - Unix datagram generator will not longer 'shear' blocks across datagrams.
+- Procfs generator now generates `pid/stat` files with `comm` in parens, matching Linux behavior.
 
 ## [0.20.10]
 ### Added