+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## [0.21.0-rc0]
 ### Added
 ### Changed
 - Maximum datagram size in bytes for unix datagram generator is 8,192.
+- Altered default cache construction from 'fixed' to 'streaming'
 - Increased maximum DogStatsD context limit from 100k to 1M
 - Procfs generator now generates process names as long as 254 characters, up from 253.
 ### Fixed